+++ src/lib/isEmpty.js
@@ -1,6 +1,6 @@
 import assertString from './util/assertString';
 
-export default function isNull(str) {
+export default function isEmpty(str) {
   assertString(str);
   return str.length === 0;
 }